+++ src/Drivers/gles3/OpenGLES3Context.cpp
@@ -46,7 +46,8 @@ void OpenGLES3Context::Init()
     TB_CORE_INFO("OpenGL Info:");
     TB_CORE_INFO("  Vendor: {0}", GLES::gl->GetString(GL_VENDOR));
     TB_CORE_INFO("  Renderer: {0}", GLES::gl->GetString(GL_RENDERER));
-    TB_CORE_INFO("  Version: {0}", GLES::gl->GetString(GL_VERSION));
+    TB_CORE_INFO("  Version: OpenGL ES 3.0");
+    TB_CORE_INFO("  Max Supported Version: {0}", GLES::gl->GetString(GL_VERSION));
 
     std::string version = reinterpret_cast<const char*>(GLES::gl->GetString(GL_VERSION));
     size_t dotPosition = version.find('.');