+++ src/Platforms/Linux/LinuxWindow.h
@@ -23,9 +23,10 @@ public:
 
     void SetVSync(bool enabled) override;
     bool IsVSync() const override;
-
     void SetResizable(bool enabled) override;
     bool IsResizable() const override;
+    void SetFullscreen(uint8_t mode) override;
+    uint8_t GetFullscreenMode() const override;
 
     void SetMinSize(uint32_t minWidth, uint32_t minHeight) override;
 
@@ -43,7 +44,8 @@ private:
         std::string Title;
         unsigned int Width, Height;
         unsigned int MinWidth, MinHeight;
-        bool Resizeable;
+        uint8_t FullscreenMode;
+        bool Resizable;
         bool VSync;
 
         EventCallbackFn EventCallback;