+++ changelog.md
@@ -1,5 +1,11 @@
 Version: 18-08-2024_1 Alpha
 Changes:
+- Created console. 
+- Application settings can be set at startup using tbconfig file. 
+- Added a way to switch between fullscreen modes in Window.
+
+Version: 18-08-2024_1 Alpha
+Changes:
 - Fixed GLTFloader oversight. 
 - Application is now a proper singleton. 
 - Instead of Config.h, now every config is in ApplicationSpecification